five

+++ project.py
@@ -93,7 +93,7 @@ print('test loss:',scores[0]*100)
 print('test accuracy:',scores[1]*100)
 try:
     f = open('/project/out.txt','w')
-    f.write(str(scores[1]*100))
+    f.write(str(int(scores[1]*100)))
 except:
     print(end="")
 finally:
